+++ RedPandaIDE/settingsdialog/executorproblemsetwidget.cpp
@@ -54,7 +54,7 @@ void ExecutorProblemSetWidget::doSave()
     pSettings->executor().setCaseEditorFontName(ui->cbFont->currentFont().family());
     pSettings->executor().setCaseEditorFontOnlyMonospaced(ui->chkOnlyMonospaced->isChecked());
     pSettings->executor().setCaseEditorFontSize(ui->spinFontSize->value());
-    pSettings->executor().setEnableCaseTimeout(ui->grpEnableTimeout->isEnabled());
+    pSettings->executor().setEnableCaseTimeout(ui->grpEnableTimeout->isChecked());
     pSettings->executor().setCaseTimeout(ui->spinCaseTimeout->value());
     pSettings->executor().save();
     pMainWindow->applySettings();